# backend/src/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Ventrix API",
    "version": "1.0.0",
    "description": "API de Ventrix POS y Ventrix Contable. Generada automáticamente desde src/routes (`npm run docs:openapi`), así que refleja siempre las rutas, roles y validaciones que hay en el código.\n\nLa sesión usa un access token corto (cabecera `Authorization: Bearer`) y una cookie httpOnly de refresh. Todo es multi-tenant: el negocio sale del token, nunca del cuerpo de la petición."
  },
  "servers": [
    {
      "url": "https://api.ventrix.lat/api/v1",
      "description": "Producción"
    },
    {
      "url": "http://localhost:4000/api/v1",
      "description": "Local"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Access token de un usuario (POS o Contable)."
      },
      "sellerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Sesión del portal de vendedoras (/seller), separada de la de usuarios."
      }
    },
    "schemas": {
      "RespuestaOk": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "data": {
            "description": "Contenido de la respuesta"
          }
        }
      },
      "Error": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "error": {
            "type": "string",
            "example": "Mensaje del error"
          }
        }
      }
    }
  },
  "paths": {
...
    "/api/v1/exports/products": {
      "get": {
        "tags": [
          "export"
        ],
        "summary": "GET /api/v1/exports/products",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Operación exitosa",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RespuestaOk"
                }
              }
            }
          },
          "401": {
            "description": "Sin sesión o token vencido",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/notifications/test": {
      "post": {
        "tags": [
          "notification"
        ],
        "summary": "Envía un push de PRUEBA a los dispositivos suscritos del usuario. Sirve para que confirme, en un toque, que las notificaciones le llegan al celular.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Operación exitosa",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RespuestaOk"
                }
              }
            }
          },
          "401": {
            "description": "Sin sesión o token vencido",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/products": {
      "get": {
        "tags": [
          "product"
        ],
        "summary": "── Standard CRUD ────────────────────────────────────────────────────────────",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Operación exitosa",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RespuestaOk"
                }
              }
            }
          },
          "401": {
            "description": "Sin sesión o token vencido",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "product"
        ],
        "summary": "POST /api/v1/products",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Operación exitosa",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RespuestaOk"
                }
              }
            }
          },
          "400": {
            "description": "Datos inválidos",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Sin sesión o token vencido",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "code": {
                    "type": "string",
                    "description": "El código es requerido"
                  },
                  "name": {
                    "type": "string",
                    "description": "El nombre es requerido"
                  },
                  "salePrice": {
                    "type": "number",
                    "description": "Precio de venta inválido",
                    "minimum": 0
                  },
                  "costPrice": {
                    "type": "number",
                    "minimum": 0
                  },
                  "stock": {
                    "type": "number",
                    "description": "Stock inválido",
                    "minimum": 0
                  },
                  "wholesalePrice": {
                    "type": "number",
                    "description": "Precio mayorista inválido",
                    "minimum": 0
                  },
                  "minStock": {
                    "type": "number",
                    "description": "Cantidad mínima inválida",
                    "minimum": 0
                  },
                  "taxRate": {
                    "type": "number",
                    "description": "IVA inválido",
                    "minimum": 0,
                    "maximum": 100
                  }
                },
                "required": [
                  "code",
                  "name",
                  "salePrice"
                ]
              }
            }
          }
        }
      }
    },
...
    "/api/v1/products/{id}/adjust-stock": {
      "patch": {
        "tags": [
          "product"
        ],
        "summary": "PATCH /api/v1/products/{id}/adjust-stock",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Operación exitosa",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RespuestaOk"
                }
              }
            }
          },
          "400": {
            "description": "Datos inválidos",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Sin sesión o token vencido",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "quantity": {
                    "type": "number",
                    "minimum": 0.001
                  },
                  "type": {
                    "type": "string",
                    "enum": [
                      "IN",
                      "OUT",
                      "ADJUSTMENT"
                    ]
                  },
                  "reason": {
                    "type": "string"
                  }
                },
                "required": [
                  "quantity",
                  "type"
                ]
              }
            }
          }
        }
      }
    },
    "/api/v1/products/{id}/duplicate": {
      "post": {
        "tags": [
          "product"
        ],
        "summary": "POST /api/v1/products/{id}/duplicate",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Operación exitosa",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RespuestaOk"
                }
              }
            }
          },
          "401": {
            "description": "Sin sesión o token vencido",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/api/v1/products/{id}/stock-by-branch": {
      "get": {
        "tags": [
          "product"
        ],
        "summary": "GET /api/v1/products/{id}/stock-by-branch",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Operación exitosa",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RespuestaOk"
                }
              }
            }
          },
          "401": {
            "description": "Sin sesión o token vencido",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
...
    "/api/v1/products/import": {
      "post": {
        "tags": [
          "product"
        ],
        "summary": "── Bulk import (supports ?dryRun=true for preview) ──────────────────────────",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Operación exitosa",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RespuestaOk"
                }
              }
            }
          },
          "401": {
            "description": "Sin sesión o token vencido",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  }
}